Don’t set “overflow: hidden” on html element
Datawrapper’s graphics appear to slightly overflow their html elements,
and setting “overflow: hidden” on the html element (as
this script does) causes the overflow within the body to be clipped,
with a visible scrollbar.

Since the overflow is less than a pixel and can be comfortably
accommodated within the margin we allocate for the graphs, the simplest
way to deal with this problem is to allow the overflow to be visible.