Add one to calculated height
I don’t know why this is necessary, but it seems to be in order to
accommodate Datawrapper’s graphics: the bounding rect height alone (even
rounded up) always leaves a small bit of overflow.

I’ve contacted Datawrapper to try and establish why this is happening:
hopefully once we’ve worked out the cause we can remove this extra
pixel.